refactor(utils): relocate shared plugin runner to src/utils/run-plugins
The Plugin contract + runPlugins first-match runner (born in the image-ref
refactor, PR #23) is about to gain a second consumer outside src/lib/mcp,
and utils must not depend on lib. Move it to src/utils/run-plugins and
point the three image-ref consumers at the new home. No behavior change.
